backend/requirements.txt
# Class360 Backend Dependencies

# Whisper for transcription
openai-whisper>=20231117

# Text-to-Speech
gtts>=2.5.0

# Video processing
ffmpeg-python>=0.2.0

# OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
pathlib2>=2.3.7
